fix failing tests, remove redundant PlotTestHarness code
